first refactor of server + client, switched protocol up, no longer sending request string with delimiters but rather simply [total request string bytesize][first file name bytesize][first file name][second file name bytesize][second file name] etc, also caught a memory leak and general touch ups

+++ source/dlinkedlist.c
@@ -10,7 +10,7 @@
 
 //public methods
 static int insert(dll  *this, int end, char *id, size_t idBytesize, dataContainer *dataContainer);
-static dataContainer   *getId(dll *this, char *id, size_t idBytesize);
+static dataContainer   *getId(dll *this, char *id, uint32_t idBytesize);
 
 //private methods
 static int       insertInitial(dll *this, dllObject *object);
@@ -49,7 +49,7 @@ dll* newDll(void)
 /*
  *  getId returns NULL if an error occurs or the item identified by id is not found in the list. On success it returns a pointer to to requested dataContainer.
  */
-static dataContainer *getId(dll *this, char *id, size_t idBytesize)
+static dataContainer *getId(dll *this, char *id, uint32_t idBytesize)
 {
   dllObject *currentObject;
